refactor: 화면 출력 텍스트 상수와 선택자 상수 파일 통합
