+++ docs/README.md
@@ -59,7 +59,7 @@
     * [x] 상품 추가 요청 시 상품명에서 HTML 태그를 제거하여 처리한다.
   * 💡 편의 사항
     * [x] 사용자가 입력란에서 엔터를 누를 시 다음 입력란 포커스 및 버튼을 클릭하여 준다.
-    * [ ] 사용자가 상품을 추가한 후 다시 상품명 입력란을 포커스하여 준다.
+    * [x] 사용자가 상품을 추가한 후 다시 상품명 입력란을 포커스하여 준다.
 * 현재 추가된 상품 현황을 상품명, 가격, 수량 순으로 출력한다.
 
 #### 자판기 보유 동전 관리 페이지 🔎
@@ -158,8 +158,7 @@
 │   ├── constants # 상수 모음
 │   │   ├── config.js # 자판기 기본 설정
 │   │   ├── constants.js # 각종 상수
-│   │   ├── display.js # 화면 출력 텍스트 상수
-│   │   └── selector.js # 선택자 상수
+│   │   └── display.js # 화면 출력 텍스트 및 선택자 상수
 │   │
 │   ├── App.js
 │   └── index.js
