+++ src/components/purchase/ReturnCoinResult.js
@@ -1,10 +1,9 @@
 import Component from '../../core/Component.js';
 import ReturnTable from '../table/ReturnTable.js';
 
-import { SELECTOR } from '../../constants/selector.js';
 import { CONSTANTS } from '../../constants/constants.js';
 import { cloneObject } from '../../utils/data-tools.js';
-import { DISPLAY } from '../../constants/display.js';
+import { DISPLAY, SELECTOR } from '../../constants/display.js';
 
 export default class ReturnCoinResult extends Component {
   init() {
