+++ src/constants/display.js
@@ -39,3 +39,18 @@ export const DISPLAY = Object.freeze({
   BUTTON_PURCHASE: '구매하기',
   BUTTON_AMOUNT_RETURN: '반환하기',
 });
+
+export const SELECTOR = Object.freeze({
+  APP: '#app',
+  PRODUCT_NAME_INPUT: '#product-name-input',
+  PRODUCT_PRICE_INPUT: '#product-price-input',
+  PRODUCT_QUANTITY_INPUT: '#product-quantity-input',
+  PRODUCT_ADD_BUTTON: '#product-add-button',
+
+  CHARGE_AMOUNT_INPUT: '#charge-input',
+  CHARGE_AMOUNT_BUTTON: '#charge-button',
+
+  PURCHASE_BUTTON: '.purchase-button',
+
+  COIN_RETURN_BUTTON: '#coin-return-button',
+});
