--- src/constants/selector.js
@@ -1,14 +1,0 @@
-export const SELECTOR = Object.freeze({
-  APP: '#app',
-  PRODUCT_NAME_INPUT: '#product-name-input',
-  PRODUCT_PRICE_INPUT: '#product-price-input',
-  PRODUCT_QUANTITY_INPUT: '#product-quantity-input',
-  PRODUCT_ADD_BUTTON: '#product-add-button',
-
-  CHARGE_AMOUNT_INPUT: '#charge-input',
-  CHARGE_AMOUNT_BUTTON: '#charge-button',
-
-  PURCHASE_BUTTON: '.purchase-button',
-
-  COIN_RETURN_BUTTON: '#coin-return-button',
-});
