+++ src/index.js
@@ -1,4 +1,4 @@
-import { SELECTOR } from './constants/selector.js';
+import { SELECTOR } from './constants/display.js';
 import { $ } from './utils/element-tools.js';
 
 import App from './App.js';
